LATE NIGHT COMMIT ROCKET ACTIVATE
This adjusts the tunnels to be fairly random and usually possible to
navigate without taking serious damage (or sometimes any damage). It
also adjusts input for falling mode only so you can hold down keys and
it moves at a steady rate, instead of repeatedly hammering keys to move.
Gold in the form of '$' is placed but doesn't do anything yet; trying to
grab gold makes the dive much more of a challenge.